+++ lib/kitchen/directions/bake_injected_exercise/bake_injected_exercise.rb
@@ -5,13 +5,15 @@ module Kitchen::Directions::BakeInjectedExercise
     alphabetical_multiparts: false,
     list_type: nil,
     add_brackets: false,
-    suppress_summary: false
+    suppress_summary: false,
+    suppress_detailed: false
   })
     options.reverse_merge!(
       alphabetical_multiparts: false,
       list_type: nil,
       add_brackets: false,
-      suppress_summary: false
+      suppress_summary: false,
+      suppress_detailed: false
     )
     V1.new.bake(exercise: exercise, options: options)
   end
@@ -73,7 +75,8 @@ module Kitchen::Directions::BakeInjectedExercise
             number: problem_letter,
             options: {
               solutions_clipboard: solutions_clipboard,
-              suppress_summary: options[:suppress_summary]
+              suppress_summary: options[:suppress_summary],
+              suppress_detailed: options[:suppress_detailed]
             }
           )
         end